feat(utxo-lib): split sign verify tests for PSBTs and legacy transactions
Split the combined test file into two separate files - one for PSBT
operations and another for legacy transaction signing and verification.

Issue: BTC-2732
